trivial: Cast guint8 to guint32 before left-shifting by 24
Shifting a guint8 value left by 24 promotes it to signed int first;
when the high bit is set this is undefined behavior per C11. Cast to
guint32 before the shift to avoid this.

+++ plugins/dell-dock/fu-dell-dock-mst.c
@@ -509,8 +509,9 @@ fu_dell_dock_mst_checksum_bank(FuDellDockMst *self,
 		return FALSE;
 	}
 	data = g_bytes_get_data(csum_bytes, NULL);
-	bank_sum = GUINT32_FROM_LE(data[0] | data[1] << 8 | /* nocheck:blocked nocheck:endian */
-				   data[2] << 16 | data[3] << 24);
+	bank_sum =
+	    GUINT32_FROM_LE(data[0] | (guint32)data[1] << 8 | /* nocheck:blocked nocheck:endian */
+			    (guint32)data[2] << 16 | (guint32)data[3] << 24);
 	g_debug("MST: Bank %u checksum: 0x%x", bank, bank_sum);
 
 	*checksum = (bank_sum == payload_sum);
